+++ mail/proton-mail-uniffi/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2024-00-00
 
+### Fixed
+
+  - Query error in `Mailbox.watchUnreadCount`.
+
 ## [0.11.33] - 2024-09-03
 
 ### Fixed